Check for null values

+++ UKParliament.CodeTest.Services/PersonService.cs
@@ -1,4 +1,5 @@
-﻿using UKParliament.CodeTest.Data;
+﻿using System.Reflection.Metadata.Ecma335;
+using UKParliament.CodeTest.Data;
 
 namespace UKParliament.CodeTest.Services
 {
@@ -13,39 +14,56 @@ namespace UKParliament.CodeTest.Services
 
         public List<Person> GetAllPeople()
         {
+            if(_context.People == null)
+            {
+                return new List<Person>();
+            }
             return _context.People.ToList();
         }
 
-        public Person GetPersonById(int id)
+        public Person? GetPersonById(int id)
         {
+            if(_context.People == null)
+            {
+                return null;
+            }
             return _context.People.FirstOrDefault(p => p.Id == id);
         }
 
         public void AddPerson(Person person)
         {
-            _context.People.Add(person);
-            _context.SaveChanges();
+            if (_context.People != null)
+            {
+                _context.People.Add(person);
+                _context.SaveChanges();
+            }
         }
 
         public void UpdatePerson(int id, Person updatedPerson)
         {
-            var existingPerson = _context.People.FirstOrDefault(p => p.Id == id);
-            if (existingPerson != null)
+            if (_context.People!= null)
             {
-                existingPerson.FirstName = updatedPerson.FirstName;
-                existingPerson.LastName = updatedPerson.LastName;
-                existingPerson.Age = updatedPerson.Age;
-                _context.SaveChanges();
+                var existingPerson = _context.People.FirstOrDefault(p => p.Id == id);
+                if (existingPerson != null)
+                {
+                    existingPerson.FirstName = updatedPerson.FirstName;
+                    existingPerson.LastName = updatedPerson.LastName;
+                    existingPerson.Age = updatedPerson.Age;
+                    _context.SaveChanges();
+                }
             }
         }
 
         public void DeletePerson(int id)
         {
-            var personToRemove = _context.People.FirstOrDefault(p => p.Id == id);
-            if (personToRemove != null)
+            if (_context.People != null)
             {
-                _context.People.Remove(personToRemove);
-                _context.SaveChanges();
+                var personToRemove = _context.People.FirstOrDefault(p => p.Id == id);
+                if (personToRemove != null)
+                {
+                    _context.People.Remove(personToRemove);
+                    _context.SaveChanges();
+                }
             }
         }
     }
